pyproject.toml
[project]
name = "streamlit-attendance-fee-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
    "supabase>=2.22.0",
]
